scraper-script/requirements.txt
google-api-python-client==2.108.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
langdetect==1.0.9
youtube-transcript-api==0.6.1
